move postMessage to sandbox directory
